Move thread pool out of CompilerService onto the InferrerService
