+++ snapi-frontend/src/main/resources/reference.conf
@@ -2,17 +2,7 @@ raw {
   check-syntax-analyzers = false
 }
 raw {
- compiler {
-    inferrer {
-        # Number of inferrer threads.
-        thread-pool-size = 8
-        # Timeout for an inference request.
-        timeout = 30s
-        # Number of entries in the inferrer cache.
-        cache-size = 100
-        # Time after which an inferrer cache entry expires if it is not accessed!
-        expiry = 30s
-    }
+  compiler {
     output-format = "hjson"
     windows-line-ending = false
     skip-phases = []
@@ -57,41 +47,49 @@ raw.rest.client {
   max-conn-per-route = 20
   max-conn-total = 100
 }
+raw.inferrer {
+  # Number of inferrer threads.
+  thread-pool-size = 8
+  # Number of entries in the inferrer cache.
+  cache-size = 100
+  # Timeout for an inference request.
+  timeout = 30s
+  # Time after which an inferrer cache entry expires if it is not accessed!
+  expiry = 30s
+}
 raw.inferrer.local {
+  # Number of files to sample while inferring wildcards or folders. Use -1 for all.
+  sample-files = 10
 
-    # Number of files to sample while inferring wildcards or folders. Use -1 for all.
-    sample-files = 10
+  encoding-detection-read-size = 1M
 
-    encoding-detection-read-size = 1M
+  csv {
+    sample-size = 10000
+    # How many lines to use to guess the separator
+    separator-sample-size = 100
+    # Weight for quoted values when choosing delimiter in csv files
+    # 0.3 is a magic number, it was chosen so test "delimiter comas in quotes" works
+    quoted-weight = 3.0
+  }
 
-    csv {
-      sample-size = 10000
-      # How many lines to use to guess the separator
-      separator-sample-size = 100
-      # Weight for quoted values when choosing delimiter in csv files
-      # 0.3 is a magic number, it was chosen so test "delimiter comas in quotes" works
-      quoted-weight = 3.0
-    }
+  hjson {
+    sample-size = 10000
+  }
 
-    hjson {
-      sample-size = 10000
-    }
+  json {
+    sample-size = 10000
+  }
 
-    json {
-      sample-size = 10000
-    }
+  text {
+    sample-size = 10000
+  }
 
-    text {
-      sample-size = 10000
-    }
+  xml {
+    sample-size = 10000
+  }
 
-    xml {
-      sample-size = 10000
-    }
-
-    # Buffered seekable input stream can be used for text-based formats.
-    use-buffered-seekable-is = true
-
+  # Buffered seekable input stream can be used for text-based formats.
+  use-buffered-seekable-is = true
 }
 raw.sources {
   rdbms {
