+++ snapi-truffle/src/main/java/raw/runtime/truffle/runtime/record/DuplicateKeyRecord.java
@@ -27,7 +27,7 @@ import com.oracle.truffle.api.object.Shape;
 import java.util.Arrays;
 import java.util.Vector;
 import raw.runtime.truffle.RawLanguage;
-import raw.utils.RecordFieldsNaming;
+import raw.client.utils.RecordFieldsNaming;
 
 @ExportLibrary(InteropLibrary.class)
 public class DuplicateKeyRecord extends DynamicObject implements TruffleObject {
